% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localF.R
\name{localF}
\alias{localF}
\title{Filtre comunidades locais}
\usage{
localF(df, meanRSA = TRUE)
}
\arguments{
\item{df}{um data frame contendo resultado das simulações}

\item{meanRSA}{se definida com TRUE, A RSA das simualçõe será usada para a média, se for FALSE, a RSA será tirar a partir da média da abundância de todas as simualações, em geral, o primeiro método pode resultar em valores diferente de 1 quando somadas a RSA, já no segundo isso não ocorre.}
}
\description{
Essa função filtra a média das comunidades locais produzidas em cada simulação.
}
