% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poolF.R
\name{poolF}
\alias{poolF}
\title{Filtre simulaçoes do Pool}
\usage{
poolF(df)
}
\arguments{
\item{df}{um data frame contendo resultado das simulações}

\item{meanRSA}{se definida com TRUE, A RSA das simualçõe será usada para a média, se for FALSE, a RSA será tirar a partir da média da abundância de todas as simualações, em geral, o primeiro método pode resultar em valores diferente de 1 quando somadas a RSA, já no segundo isso não ocorre.}
}
\description{
Essa função retira as médias da RSA e o Wi relativo para o pool de espécies das simulações.
}
